+++ aws-java-sdk-codedeploy/src/main/java/com/amazonaws/services/codedeploy/model/DeploymentCreator.java
@@ -26,7 +26,8 @@ public enum DeploymentCreator {
     CodeDeploy("CodeDeploy"),
     CodeDeployAutoUpdate("CodeDeployAutoUpdate"),
     CloudFormation("CloudFormation"),
-    CloudFormationRollback("CloudFormationRollback");
+    CloudFormationRollback("CloudFormationRollback"),
+    AutoscalingTermination("autoscalingTermination");
 
     private String value;