+++ aws-java-sdk-codedeploy/src/main/java/com/amazonaws/services/codedeploy/model/MinimumHealthyHosts.java
@@ -19,7 +19,7 @@ import com.amazonaws.protocol.ProtocolMarshaller;
 
 /**
  * <p>
- * Information about minimum healthy instance.
+ * Information about the minimum number of healthy instances.
  * </p>
  * 
  * @see <a href="http://docs.aws.amazon.com/goto/WebAPI/codedeploy-2014-10-06/MinimumHealthyHosts" target="_top">AWS API