+++ aws-java-sdk-codedeploy/src/main/java/com/amazonaws/services/codedeploy/model/transform/CreateDeploymentConfigRequestMarshaller.java
@@ -35,6 +35,8 @@ public class CreateDeploymentConfigRequestMarshaller {
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("trafficRoutingConfig").build();
     private static final MarshallingInfo<String> COMPUTEPLATFORM_BINDING = MarshallingInfo.builder(MarshallingType.STRING)
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("computePlatform").build();
+    private static final MarshallingInfo<StructuredPojo> ZONALCONFIG_BINDING = MarshallingInfo.builder(MarshallingType.STRUCTURED)
+            .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("zonalConfig").build();
 
     private static final CreateDeploymentConfigRequestMarshaller instance = new CreateDeploymentConfigRequestMarshaller();
 
@@ -56,6 +58,7 @@ public class CreateDeploymentConfigRequestMarshaller {
             protocolMarshaller.marshall(createDeploymentConfigRequest.getMinimumHealthyHosts(), MINIMUMHEALTHYHOSTS_BINDING);
             protocolMarshaller.marshall(createDeploymentConfigRequest.getTrafficRoutingConfig(), TRAFFICROUTINGCONFIG_BINDING);
             protocolMarshaller.marshall(createDeploymentConfigRequest.getComputePlatform(), COMPUTEPLATFORM_BINDING);
+            protocolMarshaller.marshall(createDeploymentConfigRequest.getZonalConfig(), ZONALCONFIG_BINDING);
         } catch (Exception e) {
             throw new SdkClientException("Unable to marshall request to JSON: " + e.getMessage(), e);
         }