+++ aws-java-sdk-codedeploy/src/main/java/com/amazonaws/services/codedeploy/model/transform/CreateDeploymentGroupRequestMarshaller.java
@@ -64,6 +64,8 @@ public class CreateDeploymentGroupRequestMarshaller {
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("onPremisesTagSet").build();
     private static final MarshallingInfo<List> TAGS_BINDING = MarshallingInfo.builder(MarshallingType.LIST).marshallLocation(MarshallLocation.PAYLOAD)
             .marshallLocationName("tags").build();
+    private static final MarshallingInfo<Boolean> TERMINATIONHOOKENABLED_BINDING = MarshallingInfo.builder(MarshallingType.BOOLEAN)
+            .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("terminationHookEnabled").build();
 
     private static final CreateDeploymentGroupRequestMarshaller instance = new CreateDeploymentGroupRequestMarshaller();
 
@@ -99,6 +101,7 @@ public class CreateDeploymentGroupRequestMarshaller {
             protocolMarshaller.marshall(createDeploymentGroupRequest.getEcsServices(), ECSSERVICES_BINDING);
             protocolMarshaller.marshall(createDeploymentGroupRequest.getOnPremisesTagSet(), ONPREMISESTAGSET_BINDING);
             protocolMarshaller.marshall(createDeploymentGroupRequest.getTags(), TAGS_BINDING);
+            protocolMarshaller.marshall(createDeploymentGroupRequest.getTerminationHookEnabled(), TERMINATIONHOOKENABLED_BINDING);
         } catch (Exception e) {
             throw new SdkClientException("Unable to marshall request to JSON: " + e.getMessage(), e);
         }