+++ aws-java-sdk-codedeploy/src/main/java/com/amazonaws/services/codedeploy/model/transform/DeploymentConfigInfoMarshaller.java
@@ -39,6 +39,8 @@ public class DeploymentConfigInfoMarshaller {
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("computePlatform").build();
     private static final MarshallingInfo<StructuredPojo> TRAFFICROUTINGCONFIG_BINDING = MarshallingInfo.builder(MarshallingType.STRUCTURED)
             .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("trafficRoutingConfig").build();
+    private static final MarshallingInfo<StructuredPojo> ZONALCONFIG_BINDING = MarshallingInfo.builder(MarshallingType.STRUCTURED)
+            .marshallLocation(MarshallLocation.PAYLOAD).marshallLocationName("zonalConfig").build();
 
     private static final DeploymentConfigInfoMarshaller instance = new DeploymentConfigInfoMarshaller();
 
@@ -62,6 +64,7 @@ public class DeploymentConfigInfoMarshaller {
             protocolMarshaller.marshall(deploymentConfigInfo.getCreateTime(), CREATETIME_BINDING);
             protocolMarshaller.marshall(deploymentConfigInfo.getComputePlatform(), COMPUTEPLATFORM_BINDING);
             protocolMarshaller.marshall(deploymentConfigInfo.getTrafficRoutingConfig(), TRAFFICROUTINGCONFIG_BINDING);
+            protocolMarshaller.marshall(deploymentConfigInfo.getZonalConfig(), ZONALCONFIG_BINDING);
         } catch (Exception e) {
             throw new SdkClientException("Unable to marshall request to JSON: " + e.getMessage(), e);
         }