+++ aws-java-sdk-codedeploy/src/main/java/com/amazonaws/services/codedeploy/model/transform/DeploymentGroupInfoJsonUnmarshaller.java
@@ -146,6 +146,10 @@ public class DeploymentGroupInfoJsonUnmarshaller implements Unmarshaller<Deploym
 
                     .unmarshall(context));
                 }
+                if (context.testExpression("terminationHookEnabled", targetDepth)) {
+                    context.nextToken();
+                    deploymentGroupInfo.setTerminationHookEnabled(context.getUnmarshaller(Boolean.class).unmarshall(context));
+                }
             } else if (token == END_ARRAY || token == END_OBJECT) {
                 if (context.getLastParsedParentElement() == null || context.getLastParsedParentElement().equals(currentParentElement)) {
                     if (context.getCurrentDepth() <= originalDepth)